2.3.0
Update logger to latest mu-ts logger. Miscellaneous fixes for build process.
