+++ .gitignore
@@ -5,6 +5,8 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+lib
+
 # Runtime data
 pids
 *.pid
