+++ .prettierrc
@@ -1,5 +1,5 @@
 {
   "singleQuote": true,
   "trailingComma": "es5",
-  "printWidth": 120
+  "printWidth": 160
 }
