+++ spec/mocks/MockAPIGatewayProxyEvent.ts
