--- src/AllowedOrigin.ts
@@ -1,6 +1,0 @@
-import { APIGatewayProxyEvent } from 'aws-lambda';
-import { HTTPAPIGatewayProxyResult } from './HTTPAPIGatewayProxyResult';
-
-export interface AllowedOrigin {
-  (event: APIGatewayProxyEvent, response: HTTPAPIGatewayProxyResult): string;
-}
