--- src/HTTPAction.ts
@@ -1,10 +1,0 @@
-/**
- * Expected HTTP Actions for endpoint routing.
- */
-export enum HTTPAction {
-  GET = 'GET',
-  PUT = 'PUT',
-  POST = 'POST',
-  PATCH = 'PATCH',
-  DELETE = 'DELETE',
-}
