--- src/HTTPBody.ts
@@ -1,6 +1,0 @@
-/**
- * An HTTP body.
- */
-export interface HTTPBody {
-  [key: string]: any;
-}
