+++ src/index.ts
@@ -1,15 +1,15 @@
-export * from './AllowedOrigin';
-export * from './cors';
-export * from './endpoint';
-export * from './endpoints';
+export * from './interfaces/AllowedOrigin';
+export * from './interfaces/EventCondition';
+export * from './interfaces/HTTPSerializer';
+export * from './interfaces/Validation';
+export * from './interfaces/ValidationHandler';
+export * from './decorators/cors';
+export * from './decorators/endpoint';
+export * from './decorators/validate';
+export * from './decorators/endpoints';
+export * from './model/HTTPAction';
+export * from './model/HTTPBody';
+export * from './model/EndpointEvent';
 export * from './EndpointRouter';
-export * from './EndpointEvent';
-export * from './EventCondition';
-export * from './HTTPAction';
 export * from './HTTPAPIGatewayProxyResult';
-export * from './HTTPBody';
-export * from './HTTPSerializer';
 export * from './JSONRedactingSerializer';
-export * from './Validation';
-export * from './validate';
-export * from './ValidationHandler';
