+++ src/interfaces/AllowedOrigin.ts
@@ -1,0 +1,6 @@
+import { APIGatewayProxyEvent } from 'aws-lambda';
+import { HTTPAPIGatewayProxyResult } from '../HTTPAPIGatewayProxyResult';
+
+export interface AllowedOrigin {
+  (event: APIGatewayProxyEvent, response: HTTPAPIGatewayProxyResult): string;
+}
