+++ src/interfaces/EventCondition.ts
@@ -1,0 +1,6 @@
+import { HTTPBody } from '../model/HTTPBody';
+import { EndpointEvent } from '../model/EndpointEvent';
+
+export interface EventCondition {
+  (body: HTTPBody | undefined, event: EndpointEvent<any>): boolean;
+}
