+++ src/interfaces/HTTPSerializer.ts
@@ -1,4 +1,4 @@
-import { HTTPBody } from './HTTPBody';
+import { HTTPBody } from '../model/HTTPBody';
 
 export interface HTTPSerializer {
   /**
