+++ src/model/HTTPAction.ts
@@ -1,0 +1,10 @@
+/**
+ * Expected HTTP Actions for endpoint routing.
+ */
+export enum HTTPAction {
+  GET = 'GET',
+  PUT = 'PUT',
+  POST = 'POST',
+  PATCH = 'PATCH',
+  DELETE = 'DELETE',
+}
