+++ src/model/HTTPBody.ts
@@ -1,0 +1,6 @@
+/**
+ * An HTTP body.
+ */
+export interface HTTPBody {
+  [key: string]: any;
+}
